+++ src/routes/Main.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import LayoutDefault from '@/components/Layout/Default';
 import LayoutTimeline from '@/components/Layout/WithTimeline';
 import About from '@/views/about';
-import Home from '@/views/home';
+import Index from '@/views/index';
 import Preview from '@/views/preview';
 
 function RoutesViews() {
@@ -11,7 +11,7 @@ function RoutesViews() {
     <BrowserRouter>
       <Routes>
         <Route path="/" element={<LayoutTimeline />}>
-          <Route index element={<Home />} />
+          <Route index element={<Index />} />
           <Route path="preview" element={<Preview />} />
         </Route>
         <Route path="/" element={<LayoutDefault />}>
